Normalize template code
